requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
facenet-pytorch==2.6.0
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python==4.12.0.88
pillow==10.2.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
